chore(clippy): fix clippy warnings - const fn, redundant clones, unnecessary cast
## Changes
- Mark SagaStepResult::success() and TestSagaExecutor::fail_at_step() as const fn
- Remove redundant clone() in TestSagaExecutor (lines 158, 256)
- Remove unnecessary usize cast in property_protocol.rs test
- Fix useless vec! to use array literal in property_tests.rs

## Verification
✅ Tests pass (179 passed)
✅ Clippy clean (0 warnings)

+++ crates/fraiseql-wire/tests/property_protocol.rs
@@ -160,7 +160,7 @@ proptest! {
         if result.is_ok() {
             let (_, consumed) = result.unwrap();
             // consumed should be tag (1) + length field (4) + body
-            prop_assert_eq!(consumed, (1 + body_len + 4) as usize);
+            prop_assert_eq!(consumed, (1 + body_len + 4));
         }
     }
